Refactor Magic Rewrite seams
